create album, delete song, skip, play etc

+++ CompanionApp/app/src/main/java/de/luh/hci/pclab/apps/music/data/DatabaseRepository.kt
@@ -3,6 +3,7 @@ package de.luh.hci.pclab.apps.music.data
 import android.content.ContentResolver
 import android.content.ContentUris
 import android.provider.MediaStore
+import androidx.lifecycle.viewModelScope
 import de.luh.hci.pclab.apps.music.model.Album
 import de.luh.hci.pclab.apps.music.model.Song
 import de.luh.hci.pclab.apps.music.model.toDomain
@@ -10,6 +11,7 @@ import de.luh.hci.pclab.apps.music.model.toEntity
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
+import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 
 class DatabaseRepository(
@@ -21,8 +23,18 @@ class DatabaseRepository(
         list -> list.map { it.toDomain() }
     }
 
-    suspend fun createAlbum(name: String): Long =
-        albumDao.insert(AlbumEntity(name = name))
+    suspend fun createAlbum(name: String, artistAl: String, songs: List<Song>): Long {
+        val albumId = albumDao.insert(AlbumEntity(
+            name = name,
+            artistAl = artistAl,
+            durationMs = songs.sumOf { it.durationMs },
+            songCount = songs.size
+        ))
+        songs.forEach { song ->
+            songDao.insert(song.copy(albumId = albumId).toEntity())
+        }
+        return albumId
+    }
 
     suspend fun deleteAlbum(album: AlbumEntity) =
         albumDao.delete(album)
@@ -33,8 +45,23 @@ class DatabaseRepository(
     suspend fun addSongToAlbum(song: Song) =
         songDao.insert(song.toEntity())
 
-    suspend fun removeSongFromAlbum(song: Song) =
-        songDao.delete(song.toEntity())
+    suspend fun removeSongFromAlbum(song: Song?) =
+        if (song != null) {
+            songDao.delete(song.toEntity())
+            val album = albumDao.getAlbumById(song.albumId)
+            if (album != null) {
+                albumDao.update(album.copy(
+                    durationMs = album.durationMs - song.durationMs,
+                    songCount = album.songCount - 1
+                ))
+            }
+            else{
+                println("Album is null")
+            }
+        }
+        else{
+            println("Song is null")
+        }
